+++ tests/app/settings.py
@@ -57,14 +57,5 @@ DJANGO_ICONS = {
             "extra_classes": "extra",
             "renderer": "django_icons.renderers.IconRenderer",
         },
-        "stacked": {
-            "name": "fa-stack",
-            "tag": "span",
-            "extra_classes": "fa-2x",
-            "content": [
-                {"name": "fas fa-square fa-stack-2x"},
-                {"name": "fab fa-twitter fa-stack-1x fa-inverse"},
-            ],
-        },
     },
 }
