+++ tests/test_material.py
@@ -3,7 +3,7 @@ from django.test import TestCase
 from .test_template_tags import render_template
 
 
-class MaterialRendererTest(TestCase):
+class MaterialTest(TestCase):
     """Test the Material Design Renderer."""
 
     def test_icons(self):
